api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.45",
    "fastapi>=0.118.0",
    "mangum>=0.19.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.37.0",
    "python-dotenv>=1.0.0",
]
